+++ src/main/java/com/project/waiter/mappers/RestaurantMapper.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Mapper
 public interface RestaurantMapper {
-    public List<RestaurantVO> getList(int num, int loc1, int loc2);
+    public List<RestaurantVO> getList(int num, double loc1, double loc2);
     public RestaurantVO get(String uuid);
     public int add(RestaurantVO restaurantVO);
 }
